+++ src/main/java/gift/exception/AlreadyExistsException.java
@@ -1,0 +1,7 @@
+package gift.exception;
+
+public class AlreadyExistsException extends RuntimeException {
+    public AlreadyExistsException(String message) {
+        super(message);
+    }
+}
