+++ src/main/java/gift/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package gift.exception;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
