--- src/main/java/gift/exception/NotFoundExceptionHandler.java
@@ -1,27 +1,0 @@
-package gift.exception;
-
-import gift.exception.category.CategoryNotFoundException;
-import gift.exception.product.ProductNotFoundException;
-import gift.exception.user.MemberNotFoundException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class NotFoundExceptionHandler {
-    @ExceptionHandler(MemberNotFoundException.class)
-    public ResponseEntity<String> handleUserNotFoundException(MemberNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(e.getMessage());
-    }
-
-    @ExceptionHandler(ProductNotFoundException.class)
-    public ResponseEntity<String> handleProductNotFoundException(ProductNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-
-    @ExceptionHandler(CategoryNotFoundException.class)
-    public ResponseEntity<String> handleCategoryNotFoundException(CategoryNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
-    }
-}
