--- src/main/java/gift/exception/category/CategoryNotFoundException.java
@@ -1,7 +1,0 @@
-package gift.exception.category;
-
-public class CategoryNotFoundException extends RuntimeException {
-    public CategoryNotFoundException(String message) {
-        super(message);
-    }
-}
