--- src/main/java/gift/exception/product/ProductAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package gift.exception.product;
-
-public class ProductAlreadyExistsException extends RuntimeException {
-    public ProductAlreadyExistsException(String message) {
-        super(message);
-    }
-}
