--- src/main/java/gift/exception/product/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package gift.exception.product;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(String message) {
-        super(message);
-    }
-}
