--- src/main/java/gift/exception/user/MemberAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package gift.exception.user;
-
-public class MemberAlreadyExistsException extends RuntimeException {
-    public MemberAlreadyExistsException(String message) {
-        super(message);
-    }
-}
